Add persistence (localStorage), fix doubling on load
This allows to save new items to localStorage, and retrieve them.
There was a bug that doubled the items on itemArray on reload, because
it was looping over all itemArray and not on existingItems as it should
have. To fix this, itemArray is first set to an empty array and the
loop is done over existingItems (which contains the saved items on
localStorage).