Render options on select tag using js.
In order to have all themes in one place, it was place on an array
called themes. It will loop each one and add it to the form.
I tried to separate saved themes on localStorage, but there is a bug in
the way that new items are added to each theme, i.e. something saved
under themeA gets pushed on themeB also. So it needs to be fixed.