Add all themes to localStorage, without overlaping
This allows to send to localStorage every new item to its correspondent
theme. It solved some issues with some undefined themes and overlapping
items that didn't belong to a given theme, e.g. some item of ThemeA went
to ThemeB and also later to ThemeB. It was solved by creating an array
with current items on each theme, and then doing JSON magic, instead of
the previous one-line assigning data to localStorage on addItem().