Fix doubling of items on reload.
Previously, this bug existed and was fixed, but as the code now allows
to have more than one entry in localStorage, it ocurred again. The
problem was that every time that the array was called on reload, it
would contain every item saved, and then will call the function addItem,
which would add every item to localStorage.
The solution that I produce was to require a new attribute or prop to
the addItem function. This is onLocalStorage, which is false for the
existingItems, and true for new items obtained from the form.

Next, I'll implement adding time data to each item, and continue with
styles.